fix the initial unit testing
